Add encoding feature and improve data export functionality
Refactor `app.py` to include encoding options for data export, update `operations/encoding.py` with a lambda function for mapping, and enhance `operations/imputation.py` for improved missing value handling.

+++ utils/quality_assessment.py
@@ -58,7 +58,7 @@ class QualityAssessor:
             
             consistency_scores.append(score)
         
-        return np.mean(consistency_scores) if consistency_scores else 100
+        return float(np.mean(consistency_scores)) if consistency_scores else 100.0
     
     def calculate_uniqueness_score(self, df: pd.DataFrame) -> float:
         """
